Address mTLS review feedback
